Make QueryResult a sealed interface and move to own file
