+++ app/src/main/java/com/hedvig/app/util/apollo/SafeQuery.kt
@@ -1,9 +1,5 @@
 package com.hedvig.app.util.apollo
 
-import arrow.core.Either
-import arrow.core.None
-import arrow.core.Option
-import arrow.core.Some
 import com.apollographql.apollo3.ApolloCall
 import com.apollographql.apollo3.api.ApolloResponse
 import com.apollographql.apollo3.api.Operation
@@ -96,33 +92,5 @@ suspend fun Call.safeGraphqlCall(): QueryResult<JSONObject> = withContext(Dispat
     QueryResult.Success(jsonObject)
   } catch (ioException: IOException) {
     QueryResult.Error.GeneralError(ioException.localizedMessage)
-  }
-}
-
-sealed class QueryResult<out T> {
-  data class Success<T>(val data: T) : QueryResult<T>()
-  sealed class Error : QueryResult<Nothing>() {
-
-    abstract val message: String?
-
-    data class NoDataError(override val message: String?) : Error()
-    data class GeneralError(override val message: String?) : Error()
-    data class QueryError(override val message: String?) : Error()
-    data class NetworkError(override val message: String?) : Error()
-  }
-
-  fun toOption(): Option<T> = when (this) {
-    is Error -> None
-    is Success -> Some(this.data)
-  }
-
-  fun toEither(): Either<Error, T> = when (this) {
-    is Error -> Either.Left(this)
-    is Success -> Either.Right(this.data)
-  }
-
-  inline fun <ErrorType> toEither(ifEmpty: (message: String?) -> ErrorType): Either<ErrorType, T> = when (this) {
-    is Error -> Either.Left(ifEmpty(message))
-    is Success -> Either.Right(this.data)
   }
 }
